converted to a get/set convention for niladics
Formerly, niladics (and pseudo-variables) were modelled as a single
function---if the argument is present it would act as a setter,
otherwise as a getter.

The new way to implement this stuff is to create functions get_X and
set_X for a symbol X.  Thus, the possibility to create niladics
functions and pseudo-variables will be available to the user, too.