fixed #43, APL command line doesn't work with dfns
In the early days of the project I implemented my idea of exposing argv
as ⍵ at the top level of APL code.  Later I abandoned it in APL scripts,
but it remained in the REPL.  After certain changes in the compiler,
this ⍵ somehow turned into a variable (as opposed to just a formal
parameter name), thus screwing up all dfns.  Unfortunately, I have no
automatic tests for the REPL, and it isn't used much, so bugs like this
tend to remain unnoticed for a long time.