made everything compile to a single apl.js file
Coffee files are included in apl.coffee by name in the correct order.

The CommonJS module system is dismantled and the top-level variables in
modules are essentially global variables.

An implementation of the Γ function was inlined, to avoid dependency on
an external library.

Doctests pass at this point, but the web demo and the CLI aren't
functional.